requirements.txt
ai2-tango[fairscale]
git+https://github.com/EleutherAI/lm-evaluation-harness.git@e18ef6e162e5b7f5d9582d998fa9c980703a3540
torchmetrics
more_itertools